# [Graph Partition] allow sharing default device context