--- a/torch/_inductor/utils.py
+++ b/torch/_inductor/utils.py
@@ -1459,6 +1459,9 @@ class IndentedBuffer:
         res.writelines(other._lines)
         return res
 
+    def contains(self, new_line: Union[DeferredLineBase, LineContext, str]) -> bool:
+        return new_line in self._lines
+
 
 class FakeIndentedBuffer(IndentedBuffer):
     def __init__(self) -> None: