@BoyuanFeng 2025-09-16
> Also any number for the overall improvement for inference?

For vllm Qwen/Qwen3-0.6B, this PR reduces latency from 0.34 seconds to 0.328 seconds, around 3.5% speedup.

Command to repro:
`vllm bench latency -O.splitting_ops=[] -O.cudagraph_mode=PIECEWISE -O.use_inductor_graph_partition=true`

We can easily see the speedup from trace. `torch/_inductor/output_code.py __call__` reduces from 5.5 ms to 5 ms.

[trace before the pr](https://interncache-all.fbcdn.net/manifold/perfetto-artifacts/tree/ui/index.html#!/?url=https://interncache-all.fbcdn.net/manifold/perfetto_internal_traces%2Ftree%2Fshared_trace%2Fboyuan_fe258c1b-a085-4671-8b2d-4e93fbc40d55_qwen3-0.6b-before-162873) vs [trace after the pr](https://interncache-all.fbcdn.net/manifold/perfetto-artifacts/tree/ui/index.html#!/?url=https://interncache-all.fbcdn.net/manifold/perfetto_internal_traces%2Ftree%2Fshared_trace%2Fboyuan_8b409d44-6971-4605-8bbd-f5a0cf9a1dc8_qwen3-0.6b-after-162873)