@shunting314 2025-09-16
> For vllm Qwen/Qwen3-0.6B, this PR reduces latency from 0.34 seconds to 0.328 seconds, around 3.5% speedup.

Want to dive a bit more on the benchmarking. How many subgraphs do we generate for Qwen/Qwen3-0.6B?  Let's say we save X device.__enter__/__exit__ each time we call the wrapper. Assuming a pair of device.__enter__/__exit__ cost Y us (~40us according to your measurement). Then the total saving for a wrapper call is (X * Y) ms.  (X * Y) / 3.5% should roughly match the latency of one inference  call.

Also what blocks us from going one step further to call device.__enter__/__exit__ only once and do multiple wrapper calls.