@BoyuanFeng 2025-09-16
> what blocks us from going one step further to call device.enter/exit only once and do multiple wrapper calls.

Yes we currently only call device.enter/exit once with multiple wrapper calls. [tlparse](https://manifold.edge.x2p.facebook.net/v0/read/tree/logs/.tmpjvLR9c/rank_0/index.html?bucketName=tlparse_reports&apiKey=tlparse_reports-key&withPayload=1&timeoutMsec=10000) and [output code](https://manifold.edge.x2p.facebook.net/v0/read/tree/logs/.tmpjvLR9c/rank_0/-_0_0_0/inductor_output_code_cyj6ffto2az4zbaqd4u2o5vk4w4rkfb57wmi4q4bagtexpgngbs7_5.txt?bucketName=tlparse_reports&apiKey=tlparse_reports-key&withPayload=1&timeoutMsec=10000). There are  29 partitions and 28 `torch.ops.vllm.unified_attention_with_output.default`s but only 1 `with torch.cuda._DeviceGuard(0):` in `def call()`.