@BoyuanFeng 2025-09-16
For Qwen3-0.6B.

Before this PR:
<img width="1072" height="522" alt="image" src="https://github.com/user-attachments/assets/f7ad449d-ba5a-4031-b6bb-c71d7d84f8e9" />

After this PR:
<img width="1071" height="405" alt="image" src="https://github.com/user-attachments/assets/68b9b886-c7a9-45ba-9308-81edc8f6186f" />

Overall, `output_code call` reduces from 5.5 ms to 5 ms. Since there are 29 subgraphs and 28 attention ops, we expect each subgraph/attention ops save 500/28=18 us. From the trace, we see around (10+27)-(4+9) = 24 us saving. Note that the saving changes a bit across attention&partition calls.

This 10% saving is larger than the end-to-end 3.5% saving. Because the model also have several other components such as update_states, prepare_inputs, etc.