@shunting314 2025-09-16
ok. didn't realize that there are so many partitions due to not capture attention kernels in cudagraphs